// src/types/api.d.ts
export interface components {
    schemas: {
// ...
        ApiErrorEnvelope: {
            error: {
                code: string;
                message: string;
                fieldErrors?: {
                    [key: string]: string;
                };
            };
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}